Fixed create-symlinks.sh and added pwndbg as a submodule

+++ create-symlinks.sh
@@ -1,10 +1,13 @@
 #!/bin/sh
 
+# for every file in the directory...
 for filename in ~/dotfiles/*
-do
+do	# thats not the README.md, this scrips or whos filename is already taken
 	if [ $filename != ~/dotfiles/create-symlinks.sh ] && \
-	   [ $filename != ~/dotfiles/README.md ]
+	   [ $filename != ~/dotfiles/README.md ] && \
+	   [ ! -e ~/.$(basename $filename) ]
 	then
+		# create a hidden symlink to it
 		ln -s $filename ~/.$(basename $filename)
 	fi
 done
